Move logic related to processing of blocks from downloaded archive into a separate file
